Iterate MgO analysis.
-Jitter correction seems to help only when events with low signal to
background are included (the signal/background threshold level is set as
a paremeter to the function powder_pattern_adhoc_jitter_get_result in
adhoc_jitter_correction.ipynb). With a threshold value of 0.2 it
actually broadens the 111 peak (however, this degradation is reduced by
doing a linear instead of cubic fit peak background).
-Vertical sample position must have changed by a large amount between
runs 913 and 916. This brings in a systematic that's difficult to
handle.